requirements.txt
yfinance>=0.2.18
pandas>=2.0.0
numpy>=1.24.0
ta>=0.10.2
matplotlib>=3.7.0
PyYAML>=6.0
APScheduler>=3.10.0